requirements.txt
alembic==1.4.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.2.3
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
flatbuffers==1.12
gevent==20.6.2
geventhttpclient==1.4.4
greenlet==0.4.16
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
locust==1.2.3
lxml==4.5.2
Mako==1.1.3
MarkupSafe==1.1.1
msgpack==1.0.0
multi-key-dict==2.0.3
packaging==20.4
pbr==5.5.0
psutil==5.7.2
psycopg2==2.8.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python-jenkins==1.7.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
rundeckrun==0.2.1
shellescape==3.8.1
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
